# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Politics & War multiplayer strategy game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register account and create nation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or nation name already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user and nation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── NATIONS ────────────────────────────────────────────────────────────────
...
  /alliances:
    get:
      operationId: listAlliances
      tags: [alliances]
      summary: List all alliances
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AllianceSummary"

    post:
      operationId: createAlliance
      tags: [alliances]
      summary: Create a new alliance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllianceInput"
      responses:
        "201":
          description: Alliance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllianceDetail"
        "400":
          description: Already in an alliance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /alliances/{allianceId}:
    get:
      operationId: getAlliance
      tags: [alliances]
      summary: Get alliance details
      parameters:
        - name: allianceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllianceDetail"
  /alliances/{allianceId}/apply:
    post:
      operationId: applyToAlliance
      tags: [alliances]
      summary: Apply to join an alliance
      parameters:
        - name: allianceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllianceApplicationInput"
      responses:
        "200":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /alliances/{allianceId}/applications/{nationId}:
    put:
      operationId: respondToApplication
      tags: [alliances]
      summary: Accept or reject a membership application
      parameters:
        - name: allianceId
          in: path
          required: true
          schema:
            type: integer
        - name: nationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationResponseInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /alliances/{allianceId}/leave:
    post:
      operationId: leaveAlliance
      tags: [alliances]
      summary: Leave an alliance
      parameters:
        - name: allianceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left alliance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── MARKET ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ─── AUTH ────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [username, password, nationName, leaderName, continent, governmentType]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          minLength: 6
        nationName:
          type: string
          minLength: 2
          maxLength: 50
        leaderName:
          type: string
          minLength: 2
          maxLength: 50
        continent:
          type: string
          enum: [North America, South America, Europe, Africa, Asia, Australia, Antarctica]
        governmentType:
          type: string
          enum: [Democracy, Republic, Monarchy, Dictatorship, Communist, Theocracy, Oligarchy, Anarchy]
        flag:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, nation]
      properties:
        user:
          $ref: "#/components/schemas/User"
        nation:
          $ref: "#/components/schemas/NationDetail"
    User:
      type: object
      required: [id, username, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
          format: date-time

    # ─── NATION ──────────────────────────────────────────────────────────────
...
    NationDetail:
      type: object
      required: [id, name, leaderName, continent, governmentType, score, cityCount, population, gdp, taxRate, taxCollectedAt, resources, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        leaderName:
          type: string
        continent:
          type: string
        governmentType:
          type: string
        score:
          type: number
        cityCount:
          type: integer
        population:
          type: integer
        gdp:
          type: number
        taxRate:
          type: number
        taxCollectedAt:
          type: ["string", "null"]
          format: date-time
        resources:
          $ref: "#/components/schemas/Resources"
        flag:
          type: ["string", "null"]
        allianceId:
          type: ["integer", "null"]
        allianceName:
          type: ["string", "null"]
        beigeUntil:
          type: ["string", "null"]
          format: date-time
        warPolicy:
          type: string
        domesticPolicy:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Resources:
      type: object
      required: [money, food, coal, oil, iron, bauxite, lead, uranium, gasoline, steel, munitions, aluminum]
      properties:
        money:
          type: number
        food:
          type: number
        coal:
          type: number
        oil:
          type: number
        iron:
          type: number
        bauxite:
          type: number
        lead:
          type: number
        uranium:
          type: number
        gasoline:
          type: number
        steel:
          type: number
        munitions:
          type: number
        aluminum:
          type: number

    # ─── CITIES ──────────────────────────────────────────────────────────────
...
    AllianceSummary:
      type: object
      required: [id, name, acronym, memberCount, score, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        acronym:
          type: string
        memberCount:
          type: integer
        score:
          type: number
        flag:
          type: ["string", "null"]
        forumUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AllianceDetail:
      type: object
      required: [id, name, acronym, members, score, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        acronym:
          type: string
        description:
          type: ["string", "null"]
        flag:
          type: ["string", "null"]
        forumUrl:
          type: ["string", "null"]
        score:
          type: number
        members:
          type: array
          items:
            $ref: "#/components/schemas/AllianceMember"
        pendingApplications:
          type: array
          items:
            $ref: "#/components/schemas/AllianceMember"
        createdAt:
          type: string
          format: date-time
    AllianceMember:
      type: object
      required: [nationId, nationName, leaderName, role, joinedAt]
      properties:
        nationId:
          type: integer
        nationName:
          type: string
        leaderName:
          type: string
        role:
          type: string
          enum: [leader, officer, member, applicant]
        score:
          type: number
        joinedAt:
          type: string
          format: date-time
    AllianceInput:
      type: object
      required: [name, acronym]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
        acronym:
          type: string
          minLength: 2
          maxLength: 8
        description:
          type: string
          maxLength: 1000
        flag:
          type: string
        forumUrl:
          type: string
    AllianceApplicationInput:
      type: object
      properties:
        message:
          type: string
          maxLength: 500
    ApplicationResponseInput:
      type: object
      required: [accepted]
      properties:
        accepted:
          type: boolean

    # ─── MARKET ──────────────────────────────────────────────────────────────
...