# lib/api-zod/src/generated/types/resources.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Politics & War multiplayer strategy game API
 * OpenAPI spec version: 0.1.0
 */

export interface Resources {
  money: number;
  food: number;
  coal: number;
  oil: number;
  iron: number;
  bauxite: number;
  lead: number;
  uranium: number;
  gasoline: number;
  steel: number;
  munitions: number;
  aluminum: number;
}